Change semantics of ProgramBuilder.reassign()
In the original Fuzzilli project, the expression "reassign a to b"
is translated to "a = b"; this is conflict with the concept in
computer science, which instead should be "b = a". In this change,
we change the original API ProgramBuilder.reassign(_:to:) to
ProgramBuilder.reassign(_:from:) to match the concept.

There is also a reassign() api in JavaScriptLifter. The semantics
of this API is not yet re-checked !